AlwaysOnTop and Cleanup
Keep forgetting to commit each change...

Basically I added a menu for 'AlwaysOnTop' and fixed some rounding
issues with the time output at the end (and in the logs).

Also cleaned up the 'progress' a bit so there is now proper indication
of what part of the array the tool is on.  It does not indicate
individual call progress, rather it indicates overall progress.  That
is, it reports on the array which has the concatenated SegID's.