Put GuiControlGet inside loop
This allows for the positions to be updated dynamically - if the tool is
on a very long run it is at the moment a bit obnoxious to try and start
it over.  This way if the points change we can update them while the
tool is running (or waiting) instead of closing it out and
reopening/rerunning etc.